Feat: Enable meme deletion from table view via Edit button and left swipe
The Edit button deletion feature was enabled by setting the navigation bar's left bar item to an instance of editButtomItem() enables the system-defined "Edit-Done" toggle button, and then overriding the view controller's setEditing method to call the setEditing method on the tableView.  The left swipe is available as a result of the tableView's data source implementing the tableView:commitEditingStyle:forRowAtIndexPath: method.